Slider Fixes
Fixed bug with not being able to obtain edge values of slider (0.0, 1.0) on mobile.

Corrected slider thumb touch position to center itself to the touch rather than to the right of the touch.

This should fix the bugs mentioned in most cases, but if the view is nested in more containers we may need to obtain the padding/margin offsets for the container view. I'll address that if there's enough demand.